refactor: complete Phase 2 - separate variable logic into modular structure
- Extract variable-specific configuration into composition-based approach
  - Restore original FieldConfig by removing variant_discriminator/variant_data fields
  - Create VariableFieldConfig with only variable-specific fields
  - Implement VariableFieldConfigExt extension trait for field analysis
  - Add VariableFieldConfigs storage to Config using composition pattern

- Extract variable parameter parsing
  - Create VariableParamsExt trait for variable_bits parameter parsing
  - Move variable_bits parsing logic to variable/params.rs
  - Update feed_params to use extension trait pattern

- Move VariantRole enum to variable module where it belongs
- Update field analysis to parse variable attributes separately
- Maintain all functionality with 106 tests passing

This completes Phase 2 of the refactoring plan, achieving ~95% preservation
of original code in core files by using extension traits and composition.
