+++ impl/src/bitfield/field_info.rs
@@ -1,12 +1,5 @@
 use super::{field_config::FieldConfig, BitfieldStruct, Config};
 
-/// Role of a field in variable-size structs
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum VariantRole {
-    Discriminator, // Field that determines which variant is active
-    Data,          // Field that contains variable-size data
-}
-
 /// Compactly stores all shared and useful information about a single `#[bitfield]` field.
 pub struct FieldInfo<'a> {
     /// The index of the field.
@@ -15,45 +8,16 @@ pub struct FieldInfo<'a> {
     pub field: &'a syn::Field,
     /// The configuration of the field.
     pub config: FieldConfig,
-    /// The variant role of this field (if any)
-    pub variant_role: Option<VariantRole>,
 }
 
 impl<'a> FieldInfo<'a> {
     /// Creates a new field info.
     pub fn new(id: usize, field: &'a syn::Field, config: FieldConfig) -> Self {
-        let variant_role = if config.is_variant_discriminator() {
-            Some(VariantRole::Discriminator)
-        } else if config.is_variant_data() {
-            Some(VariantRole::Data)
-        } else {
-            None
-        };
-
         Self {
             index: id,
             field,
             config,
-            variant_role,
         }
-    }
-
-    /// Returns true if this field is marked as a variant discriminator
-    #[allow(dead_code)]
-    pub fn is_variant_discriminator(&self) -> bool {
-        self.variant_role == Some(VariantRole::Discriminator)
-    }
-
-    /// Returns true if this field is marked as variant data
-    #[allow(dead_code)]
-    pub fn is_variant_data(&self) -> bool {
-        self.variant_role == Some(VariantRole::Data)
-    }
-
-    /// Returns true if this field is a fixed field (not part of variable bits)
-    #[allow(dead_code)]
-    pub fn is_fixed_field(&self) -> bool {
-        self.variant_role.is_none()
     }
 
     /// Returns the ident fragment for this field.
