+++ impl/src/bitfield/variable/mod.rs
@@ -4,7 +4,13 @@
 pub mod analysis;
 pub mod errors;
 pub mod expand;
+pub mod field_config;
+pub mod field_config_ext;
+pub mod params;
 
-pub use analysis::{VariableBitsAnalysis, VariableStructAnalysis};
+pub use analysis::VariableBitsAnalysis;
 pub use errors::VariableBitsError;
 pub use expand::VariableStructExpander;
+pub use field_config::VariantRole;
+pub use field_config_ext::VariableFieldConfigExt;
+pub use params::VariableParamsExt;
